requirements.txt
pandas
numpy
scikit-learn
dvc
mlflow
psycopg2-binary
argparse
jupyterlab
joblib

